refactor: Event, headquarter 패키지에 리팩토링
