+++ src/main/java/uos/uos25/headquarter/exception/EventNotFound.java
@@ -1,4 +1,4 @@
-package uos.uos25.event.exception;
+package uos.uos25.headquarter.exception;
 
 import uos.uos25.common.error.BusinessException;
 import uos.uos25.common.error.ErrorCode;
