+++ src/main/java/uos/uos25/headquarter/repository/EventRepository.java
@@ -1,9 +1,9 @@
-package uos.uos25.event.repository;
+package uos.uos25.headquarter.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import uos.uos25.event.entity.Event;
+import uos.uos25.headquarter.entity.Event;
 
 @Repository
 public interface EventRepository extends JpaRepository<Event, Long> {}
